Do not delete input files while doing punch hole gc

+++ src/blob_gc_job.cc
@@ -494,7 +494,8 @@ Status BlobGCJob::Finish() {
     mutex_->Lock();
   }
 
-  if (s.ok() && !blob_gc_->GetColumnFamilyData()->IsDropped()) {
+  if (s.ok() && !blob_gc_->GetColumnFamilyData()->IsDropped() &&
+      !blob_gc_->use_punch_hole()) {
     TEST_SYNC_POINT("BlobGCJob::Finish::BeforeDeleteInputBlobFiles");
     s = DeleteInputBlobFiles();
   }
